requirements.txt
# PyTorch est fourni par l'image de base (CUDA 12.1). Ne PAS le re-installer ici.
gradio>=4.0,<5
numpy>=1.24
pillow>=10.0
# GFPGAN optionnel (peut être désactivé via ENABLE_GFPGAN=0)
gfpgan>=1.3.8
facexlib>=0.3.0
opencv-python-headless>=4.9.0
# Dépendances utilitaires
tqdm
